Implement package-local nicknames
This is a rough first draft of support for package-local nicknames. It
does not include any tests, and I am fairly sure the use of mutexes is
mostly wrong.

This implementation does pass all of the tests defined by
trivial-package-local-nicknames, and does not break any of the existing
clisp tests.